requirements.txt
imageio
ipython
matplotlib
opencv-python
PyYAML
scikit-learn
scipy
tqdm
numpy
typing
terminaltables
Pillow
onnxruntime
setuptools
timm
easydict
attr
thop
torchsummary